Re-factoring ClientFileCache out of settings as this caused Cli to crash and makes more sense as a standalone file
